# pkg/server/sdk/api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/cloudbackups/volume": {
      "get": {
        "summary": "Return a list of backups for the specified volume",
        "operationId": "Enumerate",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiSdkCloudBackupEnumerateResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "src_volume_id",
            "description": "(optional) Source id of the volume for the request.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "cluster_id",
            "description": "(optional) Cluster id specifies the cluster for the request.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "credential_id",
            "description": "Credential id is the credential for cloud to be used for the request.",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "all",
            "description": "(optional) All indicates if the request should show cloud backups for all clusters or the current cluster.",
            "in": "query",
            "required": false,
            "type": "boolean",
            "format": "boolean"
          }
        ],
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
...
    "/v1/mountattach/unmount": {
      "post": {
        "summary": "Unmount unmounts a mounted volume in the host that the client is communicating with",
        "operationId": "Unmount",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/apiSdkVolumeUnmountResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiSdkVolumeUnmountRequest"
            }
          }
        ],
        "tags": [
          "OpenStorageMountAttach"
        ]
      }
    },
...
  },
  "definitions": {
...
    "apiSdkCloudBackupEnumerateResponse": {
      "type": "object",
      "properties": {
        "backups": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiSdkCloudBackupInfo"
          }
        }
      },
      "title": "Defines a response which lists all the backups stored by a cloud provider"
    },
...
    "apiSdkCloudBackupInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "This is the id as represented by the cloud provider"
        },
        "src_volume_id": {
          "type": "string",
          "title": "Source volumeID of the backup"
        },
        "src_volume_name": {
          "type": "string",
          "title": "Name of the sourceVolume of the backup"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time",
          "title": "Timestamp is the timestamp at which the source volume\nwas backed up to cloud"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "title": "Metadata associated with the backup"
        },
        "status": {
          "$ref": "#/definitions/apiSdkCloudBackupStatusType",
          "title": "Status indicates the status of the backup"
        }
      },
      "title": "SdkCloudBackupInfo has information about a backup stored by a cloud provider"
    },
...
    "apiSdkCloudBackupStatusType": {
      "type": "string",
      "enum": [
        "SdkCloudBackupStatusTypeUnknown",
        "SdkCloudBackupStatusTypeNotStarted",
        "SdkCloudBackupStatusTypeDone",
        "SdkCloudBackupStatusTypeAborted",
        "SdkCloudBackupStatusTypePaused",
        "SdkCloudBackupStatusTypeStopped",
        "SdkCloudBackupStatusTypeActive",
        "SdkCloudBackupStatusTypeFailed"
      ],
      "default": "SdkCloudBackupStatusTypeUnknown",
      "description": "- SdkCloudBackupStatusTypeUnknown: Unkonwn\n - SdkCloudBackupStatusTypeNotStarted: Not started\n - SdkCloudBackupStatusTypeDone: Done\n - SdkCloudBackupStatusTypeAborted: Aborted\n - SdkCloudBackupStatusTypePaused: Paused\n - SdkCloudBackupStatusTypeStopped: Stopped\n - SdkCloudBackupStatusTypeActive: Active\n - SdkCloudBackupStatusTypeFailed: Failed",
      "title": "CloudBackup status types"
    },
...
    "apiSdkVolumeUnmountRequest": {
      "type": "object",
      "properties": {
        "volume_id": {
          "type": "string",
          "title": "Id of volume"
        },
        "mount_path": {
          "type": "string",
          "title": "MountPath for device"
        },
        "options": {
          "$ref": "#/definitions/apiSdkVolumeUnmountRequestOptions",
          "title": "Options to unmount device"
        }
      },
      "title": "Defines a request to unmount a volume on the node receiving this request"
    },
    "apiSdkVolumeUnmountRequestOptions": {
      "type": "object",
      "properties": {
        "delete_mount_path": {
          "type": "boolean",
          "format": "boolean",
          "title": "Delete the mount path on the node after unmounting"
        },
        "no_delay_before_deleting_mount_path": {
          "type": "boolean",
          "format": "boolean",
          "description": "Do not wait for a delay before deleting path.\nNormally a storage driver may delay before deleting the mount path,\nwhich may be necessary to reduce the risk of race conditions. This\nchoice will remove that delay. This value is only usable when\n`delete_mount_path` is set."
        }
      },
      "title": "Options to unmount device"
    },
    "apiSdkVolumeUnmountResponse": {
      "type": "object",
      "title": "Empty response"
    },
...
  }
}